fix(migation) :  change  type Id a int vers  bigint

+++ backand/src/main/resources/db/changelog/changes/001-create-fournisseur-table.sql
@@ -1,11 +1,11 @@
 CREATE TABLE IF NOT EXISTS   fournisseur (
-                             id  int PRIMARY KEY AUTO_INCREMENT,
-                             societe VARCHAR(200) NOT NULL,
-                             adresse VARCHAR(255),
-                             contact VARCHAR(100),
-                             email VARCHAR(150),
-                             phone VARCHAR(20),
-                             ville VARCHAR(100),
-                             ice VARCHAR(50) UNIQUE,
-                             created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
+     id  BIGINT PRIMARY KEY AUTO_INCREMENT,
+     societe VARCHAR(200) NOT NULL,
+     adresse VARCHAR(255),
+     contact VARCHAR(100),
+     email VARCHAR(150),
+     phone VARCHAR(20),
+     ville VARCHAR(100),
+     ice VARCHAR(50) UNIQUE,
+     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
