+++ backand/src/main/resources/db/changelog/changes/002-create-produit-table.sql
@@ -1,8 +1,8 @@
 CREATE TABLE IF NOT EXISTS produit (
-                         id  int PRIMARY KEY AUTO_INCREMENT,
-                         nom VARCHAR(200) NOT NULL,
-                         prix DECIMAL(10, 2) NOT NULL,
-                         description TEXT,
-                         categorie VARCHAR(50) NOT NULL,
-                         created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
+     id  BIGINT PRIMARY KEY AUTO_INCREMENT,
+     nom VARCHAR(200) NOT NULL,
+     prix DECIMAL(10, 2) NOT NULL,
+     description TEXT,
+     categorie VARCHAR(50) NOT NULL,
+     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
